Move featured links into a reusable component
