+++ apps/web/components/Organization/Slice/SliceMachine.tsx
@@ -107,6 +107,10 @@ const LatestGenericListItems = dynamic(() =>
   import('@island.is/web/components').then((mod) => mod.LatestGenericListItems),
 )
 
+const FeaturedLinksSlice = dynamic(() =>
+  import('@island.is/web/components').then((mod) => mod.FeaturedLinksSlice),
+)
+
 interface SliceMachineProps {
   slice: Slice
   namespace?: Record<string, string>
@@ -220,6 +224,8 @@ export const renderSlice = (
     case 'LatestGenericListItems': {
       return <LatestGenericListItems slice={slice} />
     }
+    case 'FeaturedLinks':
+      return <FeaturedLinksSlice slice={slice} />
     default:
       return <RichText body={[slice]} />
   }
