+++ apps/web/components/real.ts
@@ -96,3 +96,5 @@ export * from './GenericList/GenericList'
 export * from './GenericList/LatestGenericListItems/LatestGenericListItems'
 export * from './connected/ParentalLeaveCalculator/ParentalLeaveCalculator'
 export * from './Organization/OrganizationIslandFooter'
+export * from './Organization/Slice/FeaturedLinksSlice/FeaturedLinksSlice'
+export * from './FeaturedLinks/FeaturedLinks'
